Move styles to the right place
