+++ packages/manager/src/components/LandingHeader/LandingHeader.tsx
@@ -2,8 +2,18 @@ import * as React from 'react';
 
 import Grid from 'src/components/Grid';
 import Button from 'src/components/Button';
+import { makeStyles, Theme } from 'src/components/core/styles';
 import DocumentationButton from 'src/components/CMR_DocumentationButton';
 import EntityHeader, { HeaderProps } from 'src/components/EntityHeader';
+
+const useStyles = makeStyles((theme: Theme) => ({
+  button: {
+    borderRadius: 3,
+    height: 40,
+    padding: 0,
+    width: 152
+  }
+}));
 
 interface Props extends Omit<HeaderProps, 'actions'> {
   docsLink: string;
@@ -17,6 +27,7 @@ interface Props extends Omit<HeaderProps, 'actions'> {
  */
 
 export const LandingHeader: React.FC<Props> = props => {
+  const classes = useStyles();
   const { docsLink, onAddNew, title } = props;
 
   const actions = React.useMemo(
@@ -24,7 +35,11 @@ export const LandingHeader: React.FC<Props> = props => {
       <Grid container direction="row" alignItems="center" justify="center">
         {onAddNew && (
           <Grid item>
-            <Button buttonType="primary" onClick={onAddNew}>
+            <Button
+              buttonType="primary"
+              className={classes.button}
+              onClick={onAddNew}
+            >
               Create a {title}
             </Button>
           </Grid>
